==> refer/5_2_concat.py <==
# import pandas as pd
# #
# # print("已经保存好人群定向的数据，开始将三个数据进行拼接操作")
# # # 最后将三个已保存的数据进行拼接即可
# # Test_Sample_Data_time = pd.read_csv('../dataset/data/chose_time.csv')
# # Test_Sample_Data_people = pd.read_csv('../dataset/data/chose_people.csv')
# # Test_Sample_Data_train = pd.read_csv('../dataset/data/data_for_train_test.csv')
# #
# # result = pd.concat([Test_Sample_Data_train, Test_Sample_Data_time, Test_Sample_Data_people], axis=1)
# # result.to_csv('../dataset/dataset_for_train/result_for_train_all.csv', index=False)
